Add `fetch-from-biosample`
Uses the `biosample_accession` from the GenBank NDJSON file to fetch
BioSample records via Entrez. Does a very simple parsing of the
BioSample record to create a JSON of all the available attributes
for the record. Outputs all records to stdout as newline-delimited JSON
records that should be redirected to a file.